+++ includes/parser.h
@@ -6,7 +6,7 @@
 /*   By: ktomoya <ktomoya@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/10 18:14:23 by ktomoya           #+#    #+#             */
-/*   Updated: 2023/11/07 11:26:41 by ktomoya          ###   ########.fr       */
+/*   Updated: 2023/11/16 08:42:52 by ktomoya          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,9 +32,9 @@ typedef struct s_node
 	struct s_node	*left;
 	struct s_node	*right;
 	const char		*str;
-	size_t 			len;
+	size_t			len;
 	char			*expand;
-    int             expand_flag;
+	int				expand_flag;
 }	t_node;
 
 // nodeの構造体にcomanndの構造体を追加する
@@ -45,10 +45,8 @@ t_node	*command_line(t_token *tokens, int *flag);
 t_node	*cmd_suffix(t_token *tok);
 t_node	*io_file(t_token *tok, int *flag);
 t_node	*io_here(t_token *tok, int *flag);
-t_node	*io_redirects(t_token *tok, int *flag);
 t_node	*io_redirect(t_token *tok, int *flag);
 t_node	*command(t_token *tok, int *flag);
-t_node	*cmd_args(t_token *tok, int *flag);
 t_node	*cmd_arg(t_token *tok, int *flag);
 t_node	*new_node(t_node_kind kind);
 void	set_node_value(t_node *node, const char *s, size_t len);
@@ -62,5 +60,7 @@ bool	consume_type(t_type type, t_token *tok);
 void	free_token(t_token *tok);
 void	free_token_list(t_token *head);
 void	free_node_tree(t_node *node);
+void	put_syntax_error(t_token *tok);
+void	*syntax_error_null(t_token *tok);
 
 #endif
